backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.103.2"}
sqlalchemy = "^2.0.22"
pypdf2 = "^3.0.1"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.9"
openai = "^1.3.5"
yake = "^0.4.8"
modal = "^0.61.95"
llama-index = "==0.9.34"
pyvespa = "^0.39.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.3"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
